受保护的 API

+++ src/app/features/ticket/actions/delete-ticket.ts
@@ -1,14 +1,28 @@
 "use server";
 
 import { setCookieByKey } from "@/actions/cookies";
-import { fromErrorToActionState } from "@/components/form/utils/to-action-state";
+import {
+  fromErrorToActionState,
+  toActionState,
+} from "@/components/form/utils/to-action-state";
 import prisma from "@/lib/prisma";
 import { ticketsPath } from "@/paths";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
+import { isOwner } from "../../auth/utils/isowner";
+import { getAuthOrRedirect } from "../../auth/queries/get-auth-or-redirect";
 
 export const deleteTicket = async (id: string) => {
+  const { user } = await getAuthOrRedirect();
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState("ERROR", "Not authorized");
+    }
     await prisma.ticket.delete({
       where: { id },
     });
